Move the DefaultInstructorCourse to shared folder and refactor import statements in InstructorCourses.jsx
